Move "safe" functions to form-utils
