+++ forms-shared/src/tax-form/mapping/shared/oddiel2Shared.ts
@@ -1,7 +1,8 @@
 import { DanZPozemkovPriznania, TaxFormData } from '../../types'
-import { parseDateFieldDate, safeArray, safeBoolean, safeNumber, safeString } from './functions'
+import { parseDateFieldDate } from './functions'
 import { oddielBaseShared } from './oddielBaseShared'
 import { calculateTaxCalculatorFormula } from '../../calculators'
+import { safeArray, safeBoolean, safeNumber, safeString } from '../../../form-utils/safeData'
 
 const mapPriznanie = (data: TaxFormData, priznanie: DanZPozemkovPriznania) => {
   const pozemky = safeArray(priznanie.pozemky).slice(0, 17)
