+++ forms-shared/src/tax-form/mapping/shared/oddiel3JedenUcelShared.ts
@@ -1,7 +1,8 @@
 import { DanZoStaviebJedenUcelPriznania, TaxFormData } from '../../types'
-import { parseDateFieldDate, safeArray, safeBoolean, safeNumber, safeString } from './functions'
+import { parseDateFieldDate } from './functions'
 import { oddielBaseShared } from './oddielBaseShared'
 import { calculateTaxCalculatorFormula } from '../../calculators'
+import { safeArray, safeBoolean, safeNumber, safeString } from '../../../form-utils/safeData'
 
 const mapPriznanie = (data: TaxFormData, priznanie: DanZoStaviebJedenUcelPriznania) => {
   const pouzitKalkulacku =
